requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.13
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
